small fixes

+++ ProgrammerGame/Assets/_Game/Scripts/Common/GameDataPresenter.cs
@@ -38,11 +38,11 @@ namespace _Game.Common
                 .Where(x => x.State == ProjectState.Active)
                 .Sum(x => x.Level / 500f) / 9f * Settings.Instance.BoostForResetBaseValue;
 
-        public long GetRewardForLevel() => 
+        public double GetRewardForLevel() => 
             IncomePerSec * Settings.Instance.TimeForLevelReward;
 
         public int CalculateLevel() => 
-            (int) Mathf.Log10(GameData.Instance.PersistentData.TotalEarnedMoney);
+            (int) Mathf.Log10((float) GameData.Instance.PersistentData.TotalEarnedMoney);
 
         public bool CanBuyNewRoom()
         {
