+++ ProgrammerGame/Assets/_Game/Scripts/Configs/PriceSettingsScriptable.cs
@@ -5,9 +5,9 @@ namespace _Game.Configs
     [CreateAssetMenu(fileName = "Price settings", menuName = "Game/Price settings", order = 0)]
     public class PriceSettingsScriptable : ScriptableObject
     {
-        [SerializeField] private long[] _prices;
+        [SerializeField] private double[] _prices;
 
-        public long GetPrice(int level) => 
+        public double GetPrice(int level) => 
             _prices[level];
     }
 }
