+++ ProgrammerGame/Assets/_Game/Scripts/Data/PersistentData.cs
@@ -5,7 +5,7 @@ namespace _Game.Data
     public class PersistentData : ISavableData
     {
         public int Level;
-        public long TotalEarnedMoney;
+        public double TotalEarnedMoney;
         public float MainBoost = 1f;
 
         public string Key => "Persistent";
