+++ ProgrammerGame/Assets/_Game/Scripts/Data/SavableData.cs
@@ -6,7 +6,7 @@ namespace _Game.Data
     [Serializable]
     public class SavableData
     {
-        public long MoneyCount = 0;
+        public double MoneyCount = 0;
         public List<ProjectData> Projects = new List<ProjectData>();
         public List<string> AutoRunnedProjects = new List<string>();
         public List<UpgradeData> Upgrades = new List<UpgradeData>();
